+++ vals/string.go
@@ -3,12 +3,12 @@ package vals
 type String interface {
 	Value
 
-	Append(string) int
+	Append(string) int64
 	Get() string
-	GetRange(int, int) string
+	GetRange(int64, int64) string
 	GetSet(string) string
 	Set(string)
-	StrLen() int
+	StrLen() int64
 }
 
 type stringval string
@@ -17,17 +17,17 @@ func (s *stringval) Type() string {
 	return "string"
 }
 
-func (s *stringval) Append(v string) int {
+func (s *stringval) Append(v string) int64 {
 	*s += stringval(v)
-	return len(*s)
+	return int64(len(*s))
 }
 
 func (s *stringval) Get() string {
 	return string(*s)
 }
 
-func (s *stringval) GetRange(start, end int) string {
-	l := len(*s)
+func (s *stringval) GetRange(start, end int64) string {
+	l := int64(len(*s))
 	if start < 0 {
 		start = l + start
 		if start < 0 {
@@ -59,8 +59,8 @@ func (s *stringval) Set(v string) {
 	*s = stringval(v)
 }
 
-func (s *stringval) StrLen() int {
-	return len(*s)
+func (s *stringval) StrLen() int64 {
+	return int64(len(*s))
 }
 
 func NewString() String {
